test: JSR310アダプタが標準機能に取り込まれたため、テストコードがJSR310アダプタに依存しないように修正
