+++ src/test/java/nablarch/common/dao/UniversalDaoTest.java
@@ -8,9 +8,7 @@ import nablarch.common.dao.DaoTestHelper.Address;
 import nablarch.common.dao.DaoTestHelper.SqlFunctionResult;
 import nablarch.common.dao.DaoTestHelper.Users;
 import nablarch.common.dao.UniversalDao.Transaction;
-import nablarch.common.dao.entity.DatePkTable;
-import nablarch.common.dao.entity.IdentityColumnEntity;
-import nablarch.common.dao.entity.TimestampPkTable;
+import nablarch.common.dao.entity.*;
 import nablarch.common.idgenerator.IdGenerator;
 import nablarch.core.db.DbExecutionContext;
 import nablarch.core.db.connection.BasicDbConnection;
@@ -28,6 +26,7 @@ import nablarch.test.support.db.helper.DatabaseTestRunner;
 import nablarch.test.support.db.helper.TargetDb;
 import nablarch.test.support.db.helper.VariousDbTestHelper;
 import nablarch.test.support.reflection.ReflectionUtil;
+import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.hamcrest.beans.HasPropertyWithValue;
 import org.junit.After;
@@ -39,6 +38,8 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 
 import java.math.BigDecimal;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -934,6 +935,44 @@ public class UniversalDaoTest {
     }
 
     /**
+     * Date and Time API（JSR-310）型（{@link LocalDate}, {@link LocalDateTime}）のフィールドを持つEntityのデータを登録できること（SQLServer以外）。
+     */
+    @Test
+    @TargetDb(exclude = TargetDb.Db.SQL_SERVER)
+    public void test_insertJsr310Column() {
+        VariousDbTestHelper.createTable(Jsr310Column.class);
+        final Jsr310Column entity = new Jsr310Column();
+        entity.id = 1L;
+        entity.localDate = LocalDate.parse("2015-04-01");
+        entity.localDateTime =  LocalDateTime.parse("2015-04-01T12:34:56");
+        UniversalDao.insert(entity);
+        connection.commit();
+
+        final Jsr310Column actual = VariousDbTestHelper.findById(Jsr310Column.class, entity.id);
+        MatcherAssert.assertThat(actual.localDate, is(entity.localDate));
+        MatcherAssert.assertThat(actual.localDateTime, is(entity.localDateTime));
+    }
+
+    /**
+     * Date and Time API（JSR-310）型（{@link LocalDate}, {@link LocalDateTime}）のフィールドを持つEntityのデータを登録できること（SQLServer用）。
+     */
+    @Test
+    @TargetDb(include = TargetDb.Db.SQL_SERVER)
+    public void test_insertJsr310Column_SQLServer() {
+        VariousDbTestHelper.createTable(Jsr310ColumnForSqlServer.class);
+        final Jsr310ColumnForSqlServer entity = new Jsr310ColumnForSqlServer();
+        entity.id = 1L;
+        entity.localDate = LocalDate.parse("2015-04-01");
+        entity.localDateTime =  LocalDateTime.parse("2015-04-01T12:34:56");
+        UniversalDao.insert(entity);
+        connection.commit();
+
+        final Jsr310ColumnForSqlServer actual = VariousDbTestHelper.findById(Jsr310ColumnForSqlServer.class, entity.id);
+        MatcherAssert.assertThat(actual.localDate, is(entity.localDate));
+        MatcherAssert.assertThat(actual.localDateTime, is(entity.localDateTime));
+    }
+
+    /**
      * CLOB型のカラムのデータを更新できること
      */
     @Test
@@ -972,6 +1011,55 @@ public class UniversalDaoTest {
         final TextColumn actual = VariousDbTestHelper.findById(TextColumn.class, entity.id);
         assertThat(actual.text, is(entity.text));
     }
+
+    /**
+     * Date and Time API（JSR-310）型（{@link LocalDate}, {@link LocalDateTime}）のフィールドを持つEntityのデータを更新できること（SQLServer以外）。
+     */
+    @Test
+    @TargetDb(exclude = TargetDb.Db.SQL_SERVER)
+    public void test_updateJsr310Column() {
+        VariousDbTestHelper.createTable(Jsr310Column.class);
+        final Jsr310Column entity = new Jsr310Column();
+        entity.id = 12345L;
+        entity.localDate = LocalDate.parse("2015-04-01");
+        entity.localDateTime =  LocalDateTime.parse("2015-04-01T12:34:56");
+        VariousDbTestHelper.insert(entity);
+
+        entity.localDate = LocalDate.parse("2014-03-31");
+        entity.localDateTime =  LocalDateTime.parse("2014-03-31T01:23:45");
+        UniversalDao.update(entity);
+        connection.commit();
+
+        final Jsr310Column actual = VariousDbTestHelper.findById(Jsr310Column.class, entity.id);
+        MatcherAssert.assertThat(actual.localDate, is(entity.localDate));
+        MatcherAssert.assertThat(actual.localDateTime, is(entity.localDateTime));
+
+    }
+
+    /**
+     * Date and Time API（JSR-310）型（{@link LocalDate}, {@link LocalDateTime}）のフィールドを持つEntityのデータを更新できること（SQLServer用）。
+     */
+    @Test
+    @TargetDb(include = TargetDb.Db.SQL_SERVER)
+    public void test_updateJsr310Column_SQLServer() {
+        VariousDbTestHelper.createTable(Jsr310ColumnForSqlServer.class);
+        final Jsr310ColumnForSqlServer entity = new Jsr310ColumnForSqlServer();
+        entity.id = 12345L;
+        entity.localDate = LocalDate.parse("2015-04-01");
+        entity.localDateTime =  LocalDateTime.parse("2015-04-01T12:34:56");
+        VariousDbTestHelper.insert(entity);
+
+        entity.localDate = LocalDate.parse("2014-03-31");
+        entity.localDateTime =  LocalDateTime.parse("2014-03-31T01:23:45");
+        UniversalDao.update(entity);
+        connection.commit();
+
+        final Jsr310ColumnForSqlServer actual = VariousDbTestHelper.findById(Jsr310ColumnForSqlServer.class, entity.id);
+        MatcherAssert.assertThat(actual.localDate, is(entity.localDate));
+        MatcherAssert.assertThat(actual.localDateTime, is(entity.localDateTime));
+
+    }
+
 
     /**
      * {@link UniversalDao#findById(Class, Object...)}を使用してCLOBカラムの値が取得できること。
@@ -1080,6 +1168,86 @@ public class UniversalDaoTest {
         assertThat(actual, hasSize(1));
         assertThat(actual.get(0).text, is(entity.text));
 
+    }
+
+    /**
+     * {@link UniversalDao#findBySqlFile(Class, String, Object)}を使用して{@link LocalDate}でのデータ検索ができること（SQLServer以外）。
+     */
+    @Test
+    @TargetDb(exclude = TargetDb.Db.SQL_SERVER)
+    public void test_findAllBySqlFile_localDate() {
+        VariousDbTestHelper.createTable(Jsr310Column.class);
+        final Jsr310Column entity = new Jsr310Column();
+        entity.id = 12345L;
+        entity.localDate = LocalDate.parse("2015-04-01");
+        VariousDbTestHelper.insert(entity);
+
+        // DBによってDateを保持する精度が異なるため、「指定したlocalDate以降であること」を検索条件とする
+        final EntityList<Jsr310Column> actual = UniversalDao.findAllBySqlFile(Jsr310Column.class,
+                "find_where_local_date_greater_than", new Object[] { entity.localDate.minusDays(1) });
+
+        MatcherAssert.assertThat(actual.size(), is(1));
+        MatcherAssert.assertThat(actual.get(0).localDate, is(entity.localDate));
+    }
+
+    /**
+     * {@link UniversalDao#findBySqlFile(Class, String, Object)}を使用して{@link LocalDate}でのデータ検索ができること（SQLServer用）。
+     */
+    @Test
+    @TargetDb(include = TargetDb.Db.SQL_SERVER)
+    public void test_findAllBySqlFile_localDate_SQLServer() {
+        VariousDbTestHelper.createTable(Jsr310ColumnForSqlServer.class);
+        final Jsr310ColumnForSqlServer entity = new Jsr310ColumnForSqlServer();
+        entity.id = 12345L;
+        entity.localDate = LocalDate.parse("2015-04-01");
+        VariousDbTestHelper.insert(entity);
+
+        // DBによってDateを保持する精度が異なるため、「指定したlocalDate以降であること」を検索条件とする
+        final EntityList<Jsr310ColumnForSqlServer> actual = UniversalDao.findAllBySqlFile(Jsr310ColumnForSqlServer.class,
+                "find_where_local_date_greater_than", new Object[] { entity.localDate.minusDays(1) });
+
+        MatcherAssert.assertThat(actual.size(), is(1));
+        MatcherAssert.assertThat(actual.get(0).localDate, is(entity.localDate));
+    }
+
+    /**
+     * {@link UniversalDao#findBySqlFile(Class, String, Object)}を使用して{@link LocalDateTime}でのデータ検索ができること（SQLServer以外）。
+     */
+    @Test
+    @TargetDb(exclude = TargetDb.Db.SQL_SERVER)
+    public void test_findAllBySqlFile_localDateTime() {
+        VariousDbTestHelper.createTable(Jsr310Column.class);
+        final Jsr310Column entity = new Jsr310Column();
+        entity.id = 12345L;
+        entity.localDateTime =  LocalDateTime.parse("2015-04-01T12:34:56");
+        VariousDbTestHelper.insert(entity);
+
+        // DBによってTimestampを保持する精度が異なるため、「指定したlocalDateTime以降であること」を検索条件とする
+        final EntityList<Jsr310Column> actual = UniversalDao.findAllBySqlFile(Jsr310Column.class,
+                "find_where_local_date_time_greater_than", new Object[] { entity.localDateTime.minusMinutes(1) });
+
+        MatcherAssert.assertThat(actual.size(), is(1));
+        MatcherAssert.assertThat(actual.get(0).localDateTime, is(entity.localDateTime));
+    }
+
+    /**
+     * {@link UniversalDao#findBySqlFile(Class, String, Object)}を使用して{@link LocalDateTime}でのデータ検索ができること（SQLServer用）。
+     */
+    @Test
+    @TargetDb(include = TargetDb.Db.SQL_SERVER)
+    public void test_findAllBySqlFile_localDateTime_SQLServer() {
+        VariousDbTestHelper.createTable(Jsr310ColumnForSqlServer.class);
+        final Jsr310ColumnForSqlServer entity = new Jsr310ColumnForSqlServer();
+        entity.id = 12345L;
+        entity.localDateTime =  LocalDateTime.parse("2015-04-01T12:34:56");
+        VariousDbTestHelper.insert(entity);
+
+        // DBによってTimestampを保持する精度が異なるため、「指定したlocalDateTime以降であること」を検索条件とする
+        final EntityList<Jsr310ColumnForSqlServer> actual = UniversalDao.findAllBySqlFile(Jsr310ColumnForSqlServer.class,
+                "find_where_local_date_time_greater_than", new Object[] { entity.localDateTime.minusMinutes(1) });
+
+        MatcherAssert.assertThat(actual.size(), is(1));
+        MatcherAssert.assertThat(actual.get(0).localDateTime, is(entity.localDateTime));
     }
 
     @Test
