--- src/test/resources/create-entity-with-jsr310-test.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<component-configuration
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://tis.co.jp/nablarch/component-configuration"
-        xsi:schemaLocation="http://tis.co.jp/nablarch/component-configuration  https://nablarch.github.io/schema/component-configuration.xsd">
-
-  <import file="db-default.xml" />
-  <import file="JSR310.xml" />
-  
-</component-configuration>
